move execfile helper to server utils, cleanup
